+++ test/bigtest/interactors/check-in.js
@@ -38,6 +38,7 @@ import ClaimedReturnedModalInteractor from './claimed-returned-modal';
   selectLoanDetails = clickable('[data-test-loan-details] a');
   selectPatronDetails = clickable('[data-test-patron-details] a');
   selectItemDetails = clickable('[data-test-item-details] a');
+  selectRequestDetails = clickable('[data-test-request-details] a');
   selectFeeFineDetails = clickable('[data-test-fee-fine-details] a');
   fillOutError = text('[data-test-check-in-scan] [class^="error"]');
   barcodePresent = isPresent('[data-test-check-in-barcode]');